# By JohSpa
# Via JohSpa
* 'master' of https://github.com/ProjectChaos/Astro.git:

Conflicts:
	bin/Debug/tutorial.exe
	main.cpp
	obj/Debug/main.o
	tutorial.depend

+++ include/Planet.h
@@ -26,13 +26,13 @@ class Planet
         void setXAcc(double val){xAcc = val;}
         void setYAcc(double val){yAcc = val;}
 
-        int  getMass() {return mass;}
-        void       setMass(int val){mass = val;}
+        double  getMass() {return mass;}
+        void       setMass(double val){mass = val;}
 
         std::string name;
         Koord k;
         double earthMass;
-        long long mass;
+        double mass;
 
         double xAcc;
         double yAcc;
